tweak: delete proxy for non-Contour autogenerate ingress

+++ logging/bin/deploy_logging.sh
@@ -59,8 +59,11 @@ if [ "$AUTOGENERATE_INGRESS" == "true" ]; then
     if [ "$INGRESS_CREATE_ROOT_PROXY" == "true" ] && [ "$INGRESS_TYPE" == "contour" ]; then
         create_root_httpproxy logging "$LOG_NS"
     else
-        log_debug "Deleting [httpproxy/v4m-logging-root-proxy] if it exists"
-        kubectl -n "$LOG_NS" delete httpproxy v4m-logging-root-proxy --ignore-not-found
+        if kubectl get crd httpproxies.projectcontour.io --ignore-not-found -o name > /dev/null; then
+
+            log_debug "Deleting [httpproxy/v4m-logging-root-proxy] if it exists"
+            kubectl -n "$LOG_NS" delete httpproxy v4m-logging-root-proxy --ignore-not-found
+        fi
     fi
 else
     if kubectl get crd httpproxies.projectcontour.io --ignore-not-found -o name > /dev/null; then
